+++ unique_toolkit/unique_toolkit/chat/schemas.py
@@ -10,6 +10,8 @@ from pydantic import (
     model_validator,
 )
 
+from unique_toolkit.content.schemas import ContentReference
+
 # set config to convert camelCase to snake_case
 model_config = ConfigDict(
     alias_generator=camelize, populate_by_name=True, arbitrary_types_allowed=True
@@ -37,15 +39,6 @@ class ToolCall(BaseModel):
     function: Function
 
 
-class Reference(BaseModel):
-    model_config = model_config
-    name: str
-    url: str | None = None
-    sequence_number: int
-    source_id: str
-    source: str
-
-
 class ChatMessage(BaseModel):
     # This model should strictly meets https://github.com/Unique-AG/monorepo/blob/master/node/apps/node-chat/src/public-api/2023-12-06/dtos/message/public-message.dto.ts
     model_config = model_config
@@ -63,7 +56,7 @@ class ChatMessage(BaseModel):
     created_at: datetime | None = None
     completed_at: datetime | None = None
     updated_at: datetime | None = None
-    references: list[Reference] | None = None
+    references: list[ContentReference] | None = None
 
     # TODO make sdk return role consistently in lowercase
     # Currently needed as sdk returns role in uppercase